% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_helpers.R
\name{make_split}
\alias{make_split}
\title{Split data into training and test sets}
\usage{
make_split(data, prop = 0.7)
}
\arguments{
\item{data}{A data frame.}

\item{prop}{Proportion of data for training set (default 0.7).}
}
\value{
A list with `train` and `test` data frames.
}
\description{
Split data into training and test sets
}
\keyword{internal}
