% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_helpers.R
\name{segment_df}
\alias{segment_df}
\title{Create a data frame of test points and their k nearest neighbors}
\usage{
segment_df(train_x, test_x, ids, k)
}
\arguments{
\item{train_x}{Numeric matrix/data frame of training predictors.}

\item{test_x}{Numeric matrix/data frame of test predictors.}

\item{ids}{Indices of nearest neighbors for each test point.}

\item{k}{Number of neighbors.}
}
\value{
Data frame suitable for plotting segments.
}
\description{
Create a data frame of test points and their k nearest neighbors
}
\keyword{internal}
